Create AccessToOps* traits to simplify creating op structs
