+++ compiler/hash-typecheck/src/ops/mod.rs
@@ -2,6 +2,12 @@
 //!
 //! Code from this module is to be used while traversing and typing the AST, in order to unify
 //! types and ensure correctness.
+use self::{
+    building::PrimitiveBuilder, reader::PrimitiveReader, substitute::Substituter, typing::Typer,
+    unify::Unifier, validate::Validator,
+};
+use crate::storage::{primitives::ScopeId, AccessToStorage, AccessToStorageMut};
+
 pub mod building;
 pub mod params;
 pub mod reader;
@@ -10,3 +16,52 @@ pub mod substitute;
 pub mod typing;
 pub mod unify;
 pub mod validate;
+
+/// Trait to access various structures that can perform typechecking queries,
+/// by a reference to a [StorageRef](crate::storage::StorageRef).
+pub trait AccessToOps: AccessToStorage {
+    /// Create an instance ofa [PrimitiveReader].
+    fn reader(&self) -> PrimitiveReader {
+        PrimitiveReader::new(self.global_storage())
+    }
+}
+
+impl<T: AccessToStorage> AccessToOps for T {}
+
+/// Trait to access various structures that can perform typechecking operations,
+/// by a reference to a [StorageRefMut](crate::storage::StorageRefMut).
+pub trait AccessToOpsMut: AccessToStorageMut {
+    /// Create an instance of [PrimitiveBuilder] from the global storage.
+    fn builder(&mut self) -> PrimitiveBuilder {
+        PrimitiveBuilder::new(self.global_storage_mut())
+    }
+
+    /// Create an instance of [PrimitiveBuilder] from the global storage, with the given scope.
+    ///
+    /// See [PrimitiveBuilder] docs for more information.
+    fn builder_with_scope(&mut self, scope: ScopeId) -> PrimitiveBuilder {
+        PrimitiveBuilder::new_with_scope(self.global_storage_mut(), scope)
+    }
+
+    /// Create an instance of [Unifier].
+    fn unifier(&mut self) -> Unifier {
+        Unifier::new(self.storages_mut())
+    }
+
+    /// Create an instance of [Substituter].
+    fn substituter(&mut self) -> Substituter {
+        Substituter::new(self.storages_mut())
+    }
+
+    /// Create an instance of [Typer].
+    fn typer(&mut self) -> Typer {
+        Typer::new(self.storages_mut())
+    }
+
+    /// Create an instance of [Validator].
+    fn validator(&mut self) -> Validator {
+        Validator::new(self.storages_mut())
+    }
+}
+
+impl<T: AccessToStorageMut> AccessToOpsMut for T {}
