+++ compiler/hash-typecheck/src/storage/mod.rs
@@ -7,10 +7,6 @@
 //! type definition will be in [GlobalStorage] because it can be accessed from any file (with the
 //! appropriate import).
 
-use hash_source::SourceId;
-
-use crate::ops::{building::PrimitiveBuilder, reader::PrimitiveReader};
-
 use self::{
     core::CoreDefs,
     mods::ModDefStore,
@@ -22,6 +18,7 @@ use self::{
     terms::TermStore,
     trts::TrtDefStore,
 };
+use hash_source::SourceId;
 
 pub mod core;
 pub mod mods;
@@ -119,11 +116,6 @@ pub struct StorageRefMut<'gs, 'ls, 'cd> {
 pub trait AccessToStorage {
     fn storages(&self) -> StorageRef;
 
-    /// Create an instance of [PrimitiveReader] from the global storage.
-    fn reader(&self) -> PrimitiveReader {
-        PrimitiveReader::new(self.global_storage())
-    }
-
     fn global_storage(&self) -> &GlobalStorage {
         self.storages().global_storage
     }
@@ -177,18 +169,6 @@ pub trait AccessToStorage {
 /// path to a [StorageRefMut] object.
 pub trait AccessToStorageMut: AccessToStorage {
     fn storages_mut(&mut self) -> StorageRefMut;
-
-    /// Create an instance of [PrimitiveBuilder] from the global storage.
-    fn builder(&mut self) -> PrimitiveBuilder {
-        PrimitiveBuilder::new(self.global_storage_mut())
-    }
-
-    /// Create an instance of [PrimitiveBuilder] from the global storage, with the given scope.
-    ///
-    /// See [PrimitiveBuilder] docs for more information.
-    fn builder_with_scope(&mut self, scope: ScopeId) -> PrimitiveBuilder {
-        PrimitiveBuilder::new_with_scope(self.global_storage_mut(), scope)
-    }
 
     fn global_storage_mut(&mut self) -> &mut GlobalStorage {
         self.storages_mut().global_storage
